fix up macOS builds
Remove the fully static option from the BUILD file and update .bazelrc
to make the libc++20 config option Linux-only. This will avoid some
static library link errors on macOS, which which doesn't have a crt0.o
or a libc++.a.

Also add some uint64_t to size_t casts needed because of the 'unsigned
long' vs. 'unsigned long long' discrepancy on macOS.

Note that this also requires the latest beta version of Xcode, because
of the "deducing this" feature of C++23 used in the ssh extension 'wire'
package.