patch memory leak in openssh configure script test (#72)
Enable P-521 elliptic key support in asan builds, by patching the test
program in the openssh configure script. The test program generates a
P-521 key but does not free it, causing an asan build of this program to
exit with a non-zero status code. This causes the configure script to
conclude incorrectly that the P-521 algorithm is not supported.

With this patch, we can remove the #ifdef OPENSSL_HAS_NISTP521 checks in
the unit tests.