patch cel-cpp for compatibility with C++23 (#135)
The upstream cel-cpp repo has merged a fix for compatibility with C++23: 
google/cel-cpp@6c51dac.

Add a patch to bazel/foreign_cc/cel-cpp.patch based on this commit. (It
doesn't apply directly because of some intervening refactoring.)

This will allow us to re-enable Envoy extensions that depend on CEL.